+++ webapp/CHANGELOG.md
@@ -1,3 +1,10 @@
+## next
+
+Features:
+    
+* Song of Ice and Fire - House Sheet
+    * first version of the House Sheet generator
+
 ## 2014-09-01
 
 Features: